chore: auth wall tidy up
