+++ client/app/components/AuthWallSpinner.vue
@@ -1,0 +1,5 @@
+<template>
+  <div class="w-[3.5em] mt-2 mx-auto">
+    <Icon name="ei:spinner-3" size="3.5em" class="animate-spin" />
+  </div>
+</template>
