+++ client/app/layouts/default.vue
@@ -1,98 +1,11 @@
 <template>
-  <div v-if="userStore.authenticated !== true" class="w-[3.5em] mt-2 mx-auto">
-    <Icon
-      name="ei:spinner-3"
-      size="3.5em"
-      class="animate-spin"
-    />
-  </div>
-  <div v-else>
-    <SidebarNav />
-    <main class="lg:pl-72">
-      <HeaderNav />
-      <div class="px-4 py-10 sm:px-6 lg:px-8 lg:py-6">
-        <slot  />
-      </div>
-    </main>
-    <OverlayModal
-      v-model:is-shown="overlayModalState.isShown"
-      :opts="overlayModalState.opts"
-      @close-ok="overlayModalState.promiseResolve"
-      @close-cancel="overlayModalState.promiseReject"
-    />
-    <NuxtSnackbar />
-  </div>
+  <AuthWallSpinner v-if="userStore.authenticated !== true" />
+  <DefaultLayoutAfterAuthWall v-else>
+    <slot  />
+  </DefaultLayoutAfterAuthWall>
 </template>
 
 <script setup lang="ts">
 import { useUserStore } from '@/stores/user'
-import { useDefaultHead } from '~/utils/head'
-import { overlayModalKey } from '../providers/providerKeys'
-import type { OverlayModal } from '../providers/providerKeys'
-
 const userStore = useUserStore()
-
-useDefaultHead()
-
-onMounted(() => onMountedScrollToHash())
-
-// OVERLAY MODAL
-
-type OverlayModalState = {
-  isShown: boolean
-  opts: Parameters<OverlayModal['openOverlayModal']>[0]
-  promiseResolve?: (value?: string | PromiseLike<string | undefined>) => void
-  promiseReject?: (reason?: any) => void
-}
-
-const overlayModalState = shallowReactive<OverlayModalState>({
-  isShown: false,
-  opts: {
-    component: undefined,
-    componentProps: undefined
-  },
-  promiseResolve: undefined,
-  promiseReject: undefined
-})
-
-provide(overlayModalKey, {
-  openOverlayModal: (opts) => {
-    overlayModalState.opts = {
-      component: opts.component,
-      componentProps: opts.componentProps ?? {},
-      mode: opts.mode ?? 'overlay'
-    }
-    return new Promise<string | undefined>((resolve, reject) => {
-      overlayModalState.promiseResolve = resolve
-      overlayModalState.promiseReject = reject
-      overlayModalState.isShown = true
-    })
-  },
-  /**
-   * Close the active overlay modal
-   */
-  closeOverlayModal: () => {
-    if (overlayModalState.promiseReject) {
-      overlayModalState.isShown = false
-      overlayModalState.promiseReject()
-    }
-  }
-})
 </script>
-
-<style>
-:root { font-family: 'Inter', sans-serif; }
-@supports (font-variation-settings: normal) {
-  :root { font-family: 'Inter var', sans-serif; }
-}
-
-body {
-  background-color: #f9fafb;
-}
-.dark body {
-  background-color: #0a0a0a;
-}
-#__nuxt {
-  min-height: 100vh
-}
-</style>
